Linux: small fixes (#111)
* Linux: fix Nix detection

Detect the `OS_ID` of Nix by probing for the presence of the `NIX_STORE`
environment variable instead of `NIX_PATH`. The latter is only set in a
`nix-shell` session but isn't when building a derivation through
`nix-build`. In contrast, the `NIX_STORE` environment variable is set in
both cases.

Signed-off-by: Vincent Haupert <mail@vincent-haupert.de>

* Linux: call binaries relative to PATH

Using an absolute path to call binaries is incompatible with
distributions which do not follow the Filesystem Hierachy Standard;
Nix is an example. Also, it is inconsistent with the rest of the code
base, let alone superfluous.

Signed-off-by: Vincent Haupert <mail@vincent-haupert.de>

* Linux: properly extract GCC major version

Calling `gcc -dumpversion` yields the full version string, e.g.,
`10.3.0`. The `build_openssl.sh` bash script uses the `-ge` number
comparison operator to check if the returned version is at least
8. This results in an error if the returned GCC version includes a patch
version; "10.3.0" isn't a valid number.

This commit fixes the version detection by only extracting the relevant
major version of GCC.

Signed-off-by: Vincent Haupert <mail@vincent-haupert.de>